fix: permissions are fine

+++ backend/routes/admin.ts
@@ -3,9 +3,11 @@ import express from "express";
 
 import * as ormL from "../orm_functions/login_user";
 import * as ormP from "../orm_functions/person";
+import * as ormSe from "../orm_functions/session_key";
 import * as rq from "../request";
 import { Responses } from "../types";
 import * as util from "../utility";
+import { errors } from "../utility";
 
 /**
  *  Attempts to list all admins in the system.
@@ -57,20 +59,39 @@ export async function modAdmin(req: express.Request): Promise<Responses.Admin> {
         .parseUpdateAdminRequest(req)
         .then((parsed) => util.isAdmin(parsed))
         .then(async (parsed) => {
-            return ormL
-                .updateLoginUser({
-                    loginUserId: parsed.data.id,
-                    isAdmin: parsed.data.isAdmin,
-                    isCoach: parsed.data.isCoach,
-                    accountStatus: parsed.data
-                        .accountStatus as account_status_enum,
-                })
-                .then((res) =>
-                    Promise.resolve({
-                        id: res.login_user_id,
-                        name: res.person.firstname + " " + res.person.lastname,
+            if (parsed.data.id !== parsed.userId) {
+                return ormL
+                    .updateLoginUser({
+                        loginUserId: parsed.data.id,
+                        isAdmin: parsed.data.isAdmin,
+                        isCoach: parsed.data.isCoach,
+                        accountStatus: parsed.data
+                            .accountStatus as account_status_enum,
                     })
-                );
+                    .then(async (res) => {
+                        console.log(res.is_admin);
+                        console.log(res.is_coach);
+                        if (!res.is_admin && !res.is_coach) {
+                            await ormL.updateLoginUser({
+                                loginUserId: res.login_user_id,
+                                isAdmin: false,
+                                isCoach: false,
+                                accountStatus: account_status_enum.DISABLED,
+                            });
+                            await ormSe.removeAllKeysForLoginUserId(
+                                res.login_user_id
+                            );
+                        }
+                        return Promise.resolve({
+                            id: res.login_user_id,
+                            name:
+                                res.person.firstname +
+                                " " +
+                                res.person.lastname,
+                        });
+                    });
+            }
+            return Promise.reject(errors.cookInvalidID());
         });
 }
 
@@ -87,11 +108,29 @@ export async function deleteAdmin(
         .parseDeleteAdminRequest(req)
         .then((parsed) => util.isAdmin(parsed))
         .then(async (parsed) => {
-            return ormL.deleteLoginUserByPersonId(parsed.data.id).then(() => {
-                return ormP
-                    .deletePersonById(parsed.data.id)
-                    .then(() => Promise.resolve({}));
-            });
+            return ormL
+                .searchLoginUserByPerson(parsed.data.id)
+                .then((logUs) => {
+                    if (
+                        logUs !== null &&
+                        logUs.login_user_id !== parsed.userId
+                    ) {
+                        console.log(logUs.is_admin);
+                        console.log(logUs.is_coach);
+                        console.log(logUs.person);
+                        console.log(logUs.login_user_id);
+                        console.log(parsed.userId);
+                        console.log(parsed.data.id);
+                        return ormL
+                            .deleteLoginUserByPersonId(parsed.data.id)
+                            .then(() => {
+                                return ormP
+                                    .deletePersonById(parsed.data.id)
+                                    .then(() => Promise.resolve({}));
+                            });
+                    }
+                    return Promise.reject(errors.cookInvalidID());
+                });
         });
 }
