+++ backend/routes/coach.ts
@@ -6,6 +6,9 @@ import * as ormP from "../orm_functions/person";
 import * as rq from "../request";
 import { Responses } from "../types";
 import * as util from "../utility";
+import * as ormSe from "../orm_functions/session_key";
+import { errors } from "../utility";
+import * as ormL from "../orm_functions/login_user";
 
 /**
  *  Attempts to list all coaches in the system.
@@ -49,7 +52,7 @@ export async function listCoaches(
  *  @returns See the API documentation. Successes are passed using
  * `Promise.resolve`, failures using `Promise.reject`.
  */
-export async function modCoach(
+/*export async function modCoach(
     req: express.Request
 ): Promise<Responses.PartialCoach> {
     return rq
@@ -64,12 +67,65 @@ export async function modCoach(
                     accountStatus: parsed.data
                         .accountStatus as account_status_enum,
                 })
-                .then((res) =>
-                    Promise.resolve({
+                .then(async (res) => {
+                    if (!res.is_admin && !res.is_coach) {
+                        await ormSe.removeAllKeysForLoginUserId(
+                            res.login_user_id
+                        );
+                    }
+                    return Promise.resolve({
                         id: res.login_user_id,
                         name: res.person.firstname + " " + res.person.lastname,
+                    });
+                });
+        });
+}*/
+
+/**
+ *  Attempts to modify a certain coach in the system.
+ *  @param req The Express.js request to extract all required data from.
+ *  @returns See the API documentation. Successes are passed using
+ * `Promise.resolve`, failures using `Promise.reject`.
+ */
+export async function modCoach(
+    req: express.Request
+): Promise<Responses.PartialCoach> {
+    return rq
+        .parseUpdateCoachRequest(req)
+        .then((parsed) => util.isAdmin(parsed))
+        .then(async (parsed) => {
+            if (parsed.data.id !== parsed.userId) {
+                return ormLU
+                    .updateLoginUser({
+                        loginUserId: parsed.data.id,
+                        isAdmin: parsed.data.isAdmin,
+                        isCoach: parsed.data.isCoach,
+                        accountStatus: parsed.data
+                            .accountStatus as account_status_enum,
                     })
-                );
+                    .then(async (res) => {
+                        if (!res.is_admin && !res.is_coach) {
+                            await ormSe.removeAllKeysForLoginUserId(
+                                res.login_user_id
+                            );
+                            await ormL.updateLoginUser({
+                                loginUserId: res.login_user_id,
+                                isAdmin: false,
+                                isCoach: false,
+                                accountStatus: account_status_enum.DISABLED,
+                            });
+                        }
+                        return Promise.resolve({
+                            id: res.login_user_id,
+                            name:
+                                res.person.firstname +
+                                " " +
+                                res.person.lastname,
+                        });
+                    });
+            } else {
+                return Promise.reject(errors.cookInvalidID());
+            }
         });
 }
 
@@ -86,11 +142,29 @@ export async function deleteCoach(
         .parseDeleteCoachRequest(req)
         .then((parsed) => util.isAdmin(parsed))
         .then(async (parsed) => {
-            return ormLU.deleteLoginUserByPersonId(parsed.data.id).then(() => {
-                return ormP
-                    .deletePersonById(parsed.data.id)
-                    .then(() => Promise.resolve({}));
-            });
+            return ormL
+                .searchLoginUserByPerson(parsed.data.id)
+                .then((logUs) => {
+                    if (
+                        logUs !== null &&
+                        logUs.login_user_id !== parsed.userId
+                    ) {
+                        console.log(logUs.is_admin);
+                        console.log(logUs.is_coach);
+                        console.log(logUs.person);
+                        console.log(logUs.login_user_id);
+                        console.log(parsed.userId);
+                        console.log(parsed.data.id);
+                        return ormL
+                            .deleteLoginUserByPersonId(parsed.data.id)
+                            .then(() => {
+                                return ormP
+                                    .deletePersonById(parsed.data.id)
+                                    .then(() => Promise.resolve({}));
+                            });
+                    }
+                    return Promise.reject(errors.cookInvalidID());
+                });
         });
 }
